Bluez: Updating out_write timeout to higher value.
The out_write timeout is set for frames to be available after
write. This value is currently set to 1sec. If it exceeds 1 sec
there will be SUSPEND issued on the stream, which will be again
moved to start when write frames are available. Updating this value
to be in sync with AudioFlinger standby timeout. This avoids frequent
SUSPEND,START switching when RF performance is poor.

CRs-Fixed: 330592
(cherry picked from commit 42a1963d34449c086c760518cae5c1d731627924)

Change-Id: Ibc60a9aa97a21175a2e2135ac6e675ec12f2c4db