Bluetooth: Create stored device from lekeys
During BlueZ initialization, we need to check the lekeys database
to see if there is any LE device which needs to be created. Currently
we only look at primary service database to figure out which LE device
should be created which is not sufficient if we have not done a primary
search on a given LE device or if the primary search failed.

CRs-fixed: 368744
(cherry picked from commit a486796ff4adf7ba8d867877a4a3a0144b0dc67a)

Change-Id: I499fd5babba72a5ab1c682e5793c968e12c3a6cb