Bluetooth: Setting authentication req to LOW for HID Mouse devices.
If a HID Mouse is 2.0 device, it doesn't support bonding proceedure.
If it is a 2.1 device, then authentication is mandatory. The current
change is to go ahead for dedicated bonding to a HID Mouse with
security level set to LOW. If the remote device features support
SSP pairing, the security level will be reset to HIGH and the
link key will be stored.

(cherry picked from commit 691e1b10c0c75c8f64e1f36764533c78ce3ab61d)

Change-Id: I586237114059bbaaeccf821bdc0e48a5d0be8a47
CRs-Fixed: 385463