BT-HID: Properly set pairstate of mouse when device created from profiles
This patch fixes the paired HID mouse not shown in paired
devices list when Aeroplane mode is turned on->off.
When Aeroplane mode is turned off, stored devices are
created from profiles. The patch sets paired values for
stored HID mouse as true. Also, same issue is seen when
device is rebooted.

Change-Id: Ic168445f72094a135808627820cbe2464466d239
CRs-Fixed: 482250