Bluetooth: A2DP: Increase the Audio Buffer size
Increase the audio buffer size to hold data for 500 ms to
ensure that audio is not dropped when the headset in SNIFF
mode takes more than 1 second to respond to START request.
Audio module has buffer for 500 ms and with this change it
can handle a delay of 1 second.

Change-Id: I0043679f56cc873a89d2a9eff1f0c9aa99ec1dd1
CRs-fixed: 482793