Bluetooth: Clean up authentication on User Cancel
This patch cleans up ongoing authentication if users presses cancels
for passkey, pincode, oob, consent, etc. Without this change
authentication was only cleaned on reception of disconnection of
ACL, which could be as long as 30 Seconds (Authentication timeout),
and during this time, it was not possible to initiate another pairing
from DUT as authentication was rejected.

Change-Id: Ic6fd4b2389a95b4385d19fbb3ceefa9170eecf38
CRs-Fixed: 524072